+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ### :house: Internal
 - `editor-common`
   - [#2622](https://github.com/wix-incubator/rich-content/pull/2622) moved textSearchInput/Separator components into ui-components folder
+- `ricos-content`
+  - [#2610](https://github.com/wix/ricos/pull/2610) tiptap converters fixed
 
 </details>
 <hr/>